+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - **Documentation**: in `html` and `markdown` format.
+
+### Fixed
+
+- Bad representation for multiline messages in stack trace, source code title and source code message.
   
 ## [0.1.0] - 2019-04-20